+++ server/src/test/java/org/elasticsearch/action/admin/indices/open/OpenIndexResponseTests.java
@@ -11,7 +11,7 @@ package org.elasticsearch.action.admin.indices.open;
 
 import org.elasticsearch.common.io.stream.Writeable;
 import org.elasticsearch.test.AbstractXContentSerializingTestCase;
-import org.elasticsearch.test.rest.ESRestTestCase;
+import org.elasticsearch.test.rest.TestResponseParsers;
 import org.elasticsearch.xcontent.ConstructingObjectParser;
 import org.elasticsearch.xcontent.XContentParser;
 
@@ -24,7 +24,7 @@ public class OpenIndexResponseTests extends AbstractXContentSerializingTestCase<
     );
 
     static {
-        ESRestTestCase.declareAcknowledgedAndShardsAcknowledgedFields(PARSER);
+        TestResponseParsers.declareAcknowledgedAndShardsAcknowledgedFields(PARSER);
     }
 
     @Override
