+++ test/framework/src/main/java/org/elasticsearch/test/rest/ESRestTestCase.java
@@ -32,14 +32,12 @@ import org.elasticsearch.TransportVersions;
 import org.elasticsearch.Version;
 import org.elasticsearch.action.admin.cluster.node.tasks.list.TransportListTasksAction;
 import org.elasticsearch.action.admin.cluster.repositories.put.PutRepositoryRequest;
-import org.elasticsearch.action.admin.cluster.settings.RestClusterGetSettingsResponse;
 import org.elasticsearch.action.admin.indices.create.CreateIndexResponse;
 import org.elasticsearch.action.fieldcaps.FieldCapabilitiesResponse;
 import org.elasticsearch.action.fieldcaps.FieldCapsUtils;
 import org.elasticsearch.action.support.broadcast.BaseBroadcastResponse;
 import org.elasticsearch.action.support.broadcast.BroadcastResponse;
 import org.elasticsearch.action.support.master.AcknowledgedResponse;
-import org.elasticsearch.action.support.master.ShardsAcknowledgedResponse;
 import org.elasticsearch.client.Request;
 import org.elasticsearch.client.RequestOptions;
 import org.elasticsearch.client.RequestOptions.Builder;
@@ -79,8 +77,6 @@ import org.elasticsearch.test.ESTestCase;
 import org.elasticsearch.xcontent.ConstructingObjectParser;
 import org.elasticsearch.xcontent.DeprecationHandler;
 import org.elasticsearch.xcontent.NamedXContentRegistry;
-import org.elasticsearch.xcontent.ObjectParser;
-import org.elasticsearch.xcontent.ParseField;
 import org.elasticsearch.xcontent.ToXContent;
 import org.elasticsearch.xcontent.XContentParser;
 import org.elasticsearch.xcontent.XContentParserConfiguration;
@@ -136,8 +132,6 @@ import static org.elasticsearch.client.RestClient.IGNORE_RESPONSE_CODES_PARAM;
 import static org.elasticsearch.cluster.ClusterState.VERSION_INTRODUCING_TRANSPORT_VERSIONS;
 import static org.elasticsearch.core.Strings.format;
 import static org.elasticsearch.test.rest.TestFeatureService.ALL_FEATURES;
-import static org.elasticsearch.xcontent.ConstructingObjectParser.constructorArg;
-import static org.elasticsearch.xcontent.ConstructingObjectParser.optionalConstructorArg;
 import static org.elasticsearch.xcontent.ToXContent.EMPTY_PARAMS;
 import static org.hamcrest.Matchers.anyOf;
 import static org.hamcrest.Matchers.containsString;
@@ -1861,7 +1855,7 @@ public abstract class ESRestTestCase extends ESTestCase {
 
         final Response response = client.performRequest(request);
         try (var parser = responseAsParser(response)) {
-            return parseCreateIndexResponse(parser);
+            return TestResponseParsers.parseCreateIndexResponse(parser);
         }
     }
 
@@ -1873,7 +1867,7 @@ public abstract class ESRestTestCase extends ESTestCase {
         Request request = new Request("DELETE", "/" + name);
         Response response = restClient.performRequest(request);
         try (var parser = responseAsParser(response)) {
-            return parseAcknowledgedResponse(parser);
+            return TestResponseParsers.parseAcknowledgedResponse(parser);
         }
     }
 
@@ -2525,97 +2519,5 @@ public abstract class ESRestTestCase extends ESTestCase {
         final var request = new Request(method.name(), endpoint);
         addXContentBody(request, body);
         return request;
-    }
-
-    private static final ConstructingObjectParser<RestClusterGetSettingsResponse, Void> REST_SETTINGS_RESPONSE_PARSER =
-        new ConstructingObjectParser<>("cluster_get_settings_response", true, a -> {
-            Settings defaultSettings = a[2] == null ? Settings.EMPTY : (Settings) a[2];
-            return new RestClusterGetSettingsResponse((Settings) a[0], (Settings) a[1], defaultSettings);
-        });
-    static {
-        REST_SETTINGS_RESPONSE_PARSER.declareObject(
-            constructorArg(),
-            (p, c) -> Settings.fromXContent(p),
-            new ParseField(RestClusterGetSettingsResponse.PERSISTENT_FIELD)
-        );
-        REST_SETTINGS_RESPONSE_PARSER.declareObject(
-            constructorArg(),
-            (p, c) -> Settings.fromXContent(p),
-            new ParseField(RestClusterGetSettingsResponse.TRANSIENT_FIELD)
-        );
-        REST_SETTINGS_RESPONSE_PARSER.declareObject(
-            optionalConstructorArg(),
-            (p, c) -> Settings.fromXContent(p),
-            new ParseField(RestClusterGetSettingsResponse.DEFAULTS_FIELD)
-        );
-    }
-
-    public static RestClusterGetSettingsResponse parseClusterSettingsResponse(XContentParser parser) {
-        return REST_SETTINGS_RESPONSE_PARSER.apply(parser, null);
-    }
-
-    private static final ParseField ACKNOWLEDGED_FIELD = new ParseField(AcknowledgedResponse.ACKNOWLEDGED_KEY);
-
-    public static <T extends AcknowledgedResponse> void declareAcknowledgedField(ConstructingObjectParser<T, Void> objectParser) {
-        objectParser.declareField(
-            constructorArg(),
-            (parser, context) -> parser.booleanValue(),
-            ACKNOWLEDGED_FIELD,
-            ObjectParser.ValueType.BOOLEAN
-        );
-    }
-
-    public static <T extends ShardsAcknowledgedResponse> void declareAcknowledgedAndShardsAcknowledgedFields(
-        ConstructingObjectParser<T, Void> objectParser
-    ) {
-        declareAcknowledgedField(objectParser);
-        objectParser.declareField(
-            constructorArg(),
-            (parser, context) -> parser.booleanValue(),
-            ShardsAcknowledgedResponse.SHARDS_ACKNOWLEDGED,
-            ObjectParser.ValueType.BOOLEAN
-        );
-    }
-
-    private static final ConstructingObjectParser<CreateIndexResponse, Void> CREATE_INDEX_RESPONSE_PARSER = new ConstructingObjectParser<>(
-        "create_index",
-        true,
-        args -> new CreateIndexResponse((boolean) args[0], (boolean) args[1], (String) args[2])
-    );
-
-    static {
-        declareAcknowledgedAndShardsAcknowledgedFields(CREATE_INDEX_RESPONSE_PARSER);
-        CREATE_INDEX_RESPONSE_PARSER.declareField(
-            constructorArg(),
-            (parser, context) -> parser.textOrNull(),
-            CreateIndexResponse.INDEX,
-            ObjectParser.ValueType.STRING
-        );
-    }
-
-    public static CreateIndexResponse parseCreateIndexResponse(XContentParser parser) {
-        return CREATE_INDEX_RESPONSE_PARSER.apply(parser, null);
-    }
-
-    /**
-     * A generic parser that simply parses the acknowledged flag
-     */
-    private static final ConstructingObjectParser<Boolean, Void> ACKNOWLEDGED_FLAG_PARSER = new ConstructingObjectParser<>(
-        "acknowledged_flag",
-        true,
-        args -> (Boolean) args[0]
-    );
-
-    static {
-        ACKNOWLEDGED_FLAG_PARSER.declareField(
-            constructorArg(),
-            (parser, context) -> parser.booleanValue(),
-            ACKNOWLEDGED_FIELD,
-            ObjectParser.ValueType.BOOLEAN
-        );
-    }
-
-    public static AcknowledgedResponse parseAcknowledgedResponse(XContentParser parser) {
-        return AcknowledgedResponse.of(ACKNOWLEDGED_FLAG_PARSER.apply(parser, null));
     }
 }
